Add initial support for Linux on IBM Z (s390x).
This commit adds pre-alpha support for s390x to SLJIT. Currently
only z12, z13 and z14 machines are likely to work due to the use of
instructions that were made available on z12. There is also no support
for multiplication overflow detection.

Generating and reading flags is likely to be slow. The flag registers
are fully emulated on s390x (since the condition code alone can't
represent two independent flag registers). This will need to be
optimized before this is ready for production use.