Initital attempt to emulate flag registers.
s390x only has one flag register, the condition code. In order to
emulate the 2 flag registers required by SLJIT we need to use a
register to hold the flags that result from an operation.

For now everything goes through this slow path. A high percentage
of the time we can use the condition code directly and we should
optimize for that.