Implement facility detection stubs.
Statically and/or dynamically detect CPU features. Static detection
is useful because GCC is often pre-configured with a recent machine
as an argument to -march. This allows the compiler to eliminate code
that will never be executed. Since this is a JIT compiler we can
also use dynamic feature detection to make use of features if they
are available at runtime.

This code is currently Linux-specific due to the use of getauxval.